fix(topic-naming): restore the naming model selector in settings
The v2 Topic Naming settings exposed only the enable switch and prompt, so the
migrated/seeded topic.naming.model_id was invisible and uneditable (#17838): an
old hidden title model kept receiving requests even after the user changed the
global default chat model.

Add a model selector bound to topic.naming.model_id in the Topic Naming panel,
filtered to chat models. Extract the shared ModelSelectorTriggerButton /
DefaultModelSelector out of ModelSettings into their own module, since
TopicNamingSettings cannot import from ModelSettings (which already imports it).
